refactor: extract text-hit and notifications from actions
Keep AppState mutations focused on session changes by moving click
hit-testing into the UI layer and agent toast/sound policy into a
dedicated notifications module.
