+++ src/app/actions.rs
@@ -12,225 +12,25 @@ use crate::layout::PaneId;
 use crate::layout::{find_in_direction, NavDirection};
 use crate::selection::Selection;
 use crate::terminal::{EffectiveStateChange, TerminalStateMutation};
+use crate::ui::{logical_cell_for_visible_cell, url_at_column};
 use crate::workspace::WorkspaceGitStatus;
 
 use super::api_helpers::pane_agent_status;
+use super::notifications::{
+    active_tab_suppresses_notifications, is_completion_transition, notification_context,
+    notification_sound_for_effective_state_change, notification_toast_for_effective_state_change,
+    sound_for_toast_kind, toast_agent_label, toast_event_text,
+};
 use super::state::{
     AgentNotificationDelivery, AppState, Mode, PaneFocusTarget, PendingAgentNotification,
     ToastKind, ToastNotification, ToastTarget,
 };
-
-fn is_background_completion_transition(prev_state: AgentState, new_state: AgentState) -> bool {
-    matches!(new_state, AgentState::Idle)
-        && matches!(prev_state, AgentState::Working | AgentState::Blocked)
-}
-
-fn is_completion_transition(change: &EffectiveStateChange) -> bool {
-    is_completion_transition_parts(
-        change.previous_state,
-        change.state,
-        change.previous_agent_label.as_deref(),
-        change.agent_label.as_deref(),
-    )
-}
 
 fn public_tab_id_for_index(ws: &crate::workspace::Workspace, tab_idx: usize) -> Option<String> {
     let tab_number = ws.public_tab_number(tab_idx)?;
     Some(crate::workspace::public_tab_id_for_number(
         &ws.id, tab_number,
     ))
-}
-
-pub fn is_completion_transition_parts(
-    previous_state: AgentState,
-    state: AgentState,
-    previous_agent_label: Option<&str>,
-    agent_label: Option<&str>,
-) -> bool {
-    is_background_completion_transition(previous_state, state)
-        || (previous_state == AgentState::Unknown
-            && state == AgentState::Idle
-            && previous_agent_label.is_some()
-            && previous_agent_label == agent_label)
-}
-
-pub fn active_tab_suppresses_notifications(
-    is_active_tab: bool,
-    outer_terminal_focus: Option<bool>,
-) -> bool {
-    is_active_tab && outer_terminal_focus != Some(false)
-}
-
-#[cfg(test)]
-pub fn notification_sound_for_state_change(
-    suppress_active_tab_notifications: bool,
-    prev_state: AgentState,
-    new_state: AgentState,
-) -> Option<crate::sound::Sound> {
-    if new_state == prev_state {
-        return None;
-    }
-
-    match new_state {
-        AgentState::Blocked => Some(crate::sound::Sound::Request),
-        AgentState::Idle
-            if is_background_completion_transition(prev_state, new_state)
-                && !suppress_active_tab_notifications =>
-        {
-            Some(crate::sound::Sound::Done)
-        }
-        _ => None,
-    }
-}
-
-pub fn notification_sound_for_state_change_with_agent_labels(
-    suppress_active_tab_notifications: bool,
-    prev_state: AgentState,
-    new_state: AgentState,
-    previous_agent_label: Option<&str>,
-    agent_label: Option<&str>,
-) -> Option<crate::sound::Sound> {
-    if new_state == prev_state {
-        return None;
-    }
-
-    match new_state {
-        AgentState::Blocked => Some(crate::sound::Sound::Request),
-        AgentState::Idle
-            if is_completion_transition_parts(
-                prev_state,
-                new_state,
-                previous_agent_label,
-                agent_label,
-            ) && !suppress_active_tab_notifications =>
-        {
-            Some(crate::sound::Sound::Done)
-        }
-        _ => None,
-    }
-}
-
-fn notification_sound_for_effective_state_change(
-    suppress_active_tab_notifications: bool,
-    change: &EffectiveStateChange,
-) -> Option<crate::sound::Sound> {
-    if change.state == change.previous_state {
-        return None;
-    }
-
-    match change.state {
-        AgentState::Blocked => Some(crate::sound::Sound::Request),
-        AgentState::Idle
-            if is_completion_transition(change) && !suppress_active_tab_notifications =>
-        {
-            Some(crate::sound::Sound::Done)
-        }
-        _ => None,
-    }
-}
-
-pub fn notification_toast_for_state_change_with_agent_labels(
-    suppress_active_tab_notifications: bool,
-    prev_state: AgentState,
-    new_state: AgentState,
-    previous_agent_label: Option<&str>,
-    agent_label: Option<&str>,
-) -> Option<ToastKind> {
-    if suppress_active_tab_notifications || new_state == prev_state {
-        return None;
-    }
-
-    match new_state {
-        AgentState::Blocked => Some(ToastKind::NeedsAttention),
-        AgentState::Idle
-            if is_completion_transition_parts(
-                prev_state,
-                new_state,
-                previous_agent_label,
-                agent_label,
-            ) =>
-        {
-            Some(ToastKind::Finished)
-        }
-        _ => None,
-    }
-}
-
-fn notification_toast_for_effective_state_change(
-    suppress_active_tab_notifications: bool,
-    change: &EffectiveStateChange,
-) -> Option<ToastKind> {
-    if suppress_active_tab_notifications || change.state == change.previous_state {
-        return None;
-    }
-
-    match change.state {
-        AgentState::Blocked => Some(ToastKind::NeedsAttention),
-        AgentState::Idle if is_completion_transition(change) => Some(ToastKind::Finished),
-        _ => None,
-    }
-}
-
-pub fn notification_toast_for_pane_state_update(
-    suppress_active_tab_notifications: bool,
-    update: &PaneStateUpdate,
-) -> Option<ToastKind> {
-    if update.suppress_completion
-        || suppress_active_tab_notifications
-        || update.state == update.previous_state
-    {
-        return None;
-    }
-
-    notification_toast_for_state_change_with_agent_labels(
-        suppress_active_tab_notifications,
-        update.previous_state,
-        update.state,
-        update.previous_agent_label.as_deref(),
-        update.agent_label.as_deref(),
-    )
-}
-
-fn toast_agent_label(agent_label: &str) -> &str {
-    agent_label
-}
-
-fn toast_event_text(kind: ToastKind) -> &'static str {
-    match kind {
-        ToastKind::NeedsAttention => "needs attention",
-        ToastKind::Finished => "finished",
-        ToastKind::UpdateInstalled => "updated",
-    }
-}
-
-fn sound_for_toast_kind(
-    kind: ToastKind,
-    suppress_active_tab_notifications: bool,
-) -> Option<crate::sound::Sound> {
-    match kind {
-        ToastKind::NeedsAttention => Some(crate::sound::Sound::Request),
-        ToastKind::Finished if !suppress_active_tab_notifications => {
-            Some(crate::sound::Sound::Done)
-        }
-        ToastKind::Finished | ToastKind::UpdateInstalled => None,
-    }
-}
-
-pub fn notification_context(
-    ws: &crate::workspace::Workspace,
-    workspace_label: &str,
-    ws_idx: usize,
-    pane_id: PaneId,
-) -> String {
-    let mut context = format!("{} · {}", workspace_label, ws_idx + 1);
-    if ws.tabs.len() > 1 {
-        if let Some(tab_idx) = ws.find_tab_index_for_pane(pane_id) {
-            if let Some(label) = ws.tab_display_name(tab_idx) {
-                context.push_str(&format!(" · {label}"));
-            }
-        }
-    }
-    context
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -1154,370 +954,6 @@ fn url_at_runtime_cell(
     url_at_column(line, logical_cell.logical_col).map(str::to_owned)
 }
 
-pub(crate) fn safe_web_url(url: &str) -> Option<&str> {
-    (url.starts_with("http://") || url.starts_with("https://")).then_some(url)
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct TextCell {
-    ch: char,
-    start_col: u16,
-    end_col: u16,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct CellSpan {
-    start: usize,
-    end: usize,
-}
-
-impl CellSpan {
-    fn contains(self, idx: usize) -> bool {
-        idx >= self.start && idx <= self.end
-    }
-
-    fn columns(self, cells: &[TextCell]) -> (u16, u16) {
-        (cells[self.start].start_col, cells[self.end].end_col)
-    }
-}
-
-/// Finds the terminal display-column bounds for the token under a double-click.
-///
-/// The algorithm first maps text to terminal cells so wide characters and
-/// zero-width marks use display columns, then prefers structured spans that
-/// users expect to copy whole (URLs and quoted paths), and finally falls back
-/// to a separator-delimited token.
-pub(crate) fn word_bounds_at_column(row: &str, col: u16) -> Option<(u16, u16)> {
-    // Map the row into display cells before doing any word-boundary work.
-    let cells = text_cells(row);
-    let clicked_idx = cell_index_at_column(&cells, col)?;
-
-    // Prefer spans that can legally include punctuation or spaces.
-    let span = url_span_at_column(&cells, clicked_idx)
-        .or_else(|| quoted_path_span_at_column(&cells, clicked_idx))
-        .or_else(|| token_span_at_column(&cells, clicked_idx))?;
-
-    // Convert the internal cell span back to inclusive terminal columns.
-    Some(span.columns(&cells))
-}
-
-pub(crate) fn url_at_column(row: &str, col: u16) -> Option<&str> {
-    let cells = text_cells(row);
-    let clicked_idx = cell_index_at_column(&cells, col)?;
-    let span = url_spans(&cells)
-        .into_iter()
-        .find(|span| span.contains(clicked_idx))?;
-    let start_byte = byte_index_for_cell(row, span.start);
-    let end_byte = byte_index_after_cell(row, span.end);
-    safe_web_url(row.get(start_byte..end_byte)?)
-}
-
-fn url_spans(cells: &[TextCell]) -> Vec<CellSpan> {
-    let mut spans = Vec::new();
-    let mut start = 0;
-    while start < cells.len() {
-        if starts_with_chars(&cells[start..], "http://")
-            || starts_with_chars(&cells[start..], "https://")
-        {
-            let mut end = start;
-            while end + 1 < cells.len() && !cells[end + 1].ch.is_whitespace() {
-                end += 1;
-            }
-            if let Some(span) = trim_url_edges(cells, CellSpan { start, end }) {
-                spans.push(span);
-            }
-            start = end + 1;
-        } else {
-            start += 1;
-        }
-    }
-    spans
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct VisibleTextCell {
-    pub(crate) byte_index: usize,
-    pub(crate) ch: char,
-    pub(crate) logical_col: u16,
-    pub(crate) screen_row: u16,
-    pub(crate) screen_col: u16,
-}
-
-pub(crate) fn visible_text_cells(text: &str, pane_width: u16) -> Vec<VisibleTextCell> {
-    if pane_width == 0 {
-        return Vec::new();
-    }
-
-    let mut cells = Vec::new();
-    let mut screen_row = 0u16;
-    let mut screen_col = 0u16;
-    let mut logical_col = 0u16;
-    let mut pending_wrap = false;
-    for (byte_index, ch) in text.char_indices() {
-        if ch == '\n' {
-            screen_row = screen_row.saturating_add(1);
-            screen_col = 0;
-            logical_col = 0;
-            pending_wrap = false;
-            continue;
-        }
-        if pending_wrap {
-            screen_row = screen_row.saturating_add(1);
-            screen_col = 0;
-            pending_wrap = false;
-        }
-
-        let width = u16::from(crate::ghostty::unicode_codepoint_width(ch as u32));
-        cells.push(VisibleTextCell {
-            byte_index,
-            ch,
-            logical_col,
-            screen_row,
-            screen_col,
-        });
-
-        logical_col = logical_col.saturating_add(width);
-        screen_col = screen_col.saturating_add(width);
-        while screen_col > pane_width {
-            screen_col -= pane_width;
-            screen_row = screen_row.saturating_add(1);
-        }
-        if width > 0 && screen_col == pane_width {
-            pending_wrap = true;
-            screen_col = pane_width.saturating_sub(1);
-        }
-    }
-    cells
-}
-
-pub(crate) fn logical_cell_for_visible_cell(
-    text: &str,
-    pane_width: u16,
-    target_row: u16,
-    target_col: u16,
-) -> Option<VisibleTextCell> {
-    visible_text_cells(text, pane_width)
-        .into_iter()
-        .find(|cell| {
-            let width = u16::from(crate::ghostty::unicode_codepoint_width(cell.ch as u32));
-            cell.screen_row == target_row
-                && if width == 0 {
-                    target_col == cell.screen_col
-                } else {
-                    target_col >= cell.screen_col
-                        && target_col < cell.screen_col.saturating_add(width)
-                }
-        })
-}
-
-fn token_span_at_column(cells: &[TextCell], clicked_idx: usize) -> Option<CellSpan> {
-    if is_word_separator(cells[clicked_idx].ch) {
-        return None;
-    }
-
-    let mut start = clicked_idx;
-    while start > 0 && !is_word_separator(cells[start - 1].ch) {
-        start -= 1;
-    }
-
-    let mut end = clicked_idx;
-    while end + 1 < cells.len() && !is_word_separator(cells[end + 1].ch) {
-        end += 1;
-    }
-
-    trim_token_edges(cells, CellSpan { start, end }).filter(|span| span.contains(clicked_idx))
-}
-
-fn text_cells(row: &str) -> Vec<TextCell> {
-    let mut next_col = 0u16;
-    row.chars()
-        .map(|ch| {
-            let width = u16::from(crate::ghostty::unicode_codepoint_width(ch as u32));
-            let start_col = if width == 0 {
-                next_col.saturating_sub(1)
-            } else {
-                next_col
-            };
-            if width > 0 {
-                next_col = next_col.saturating_add(width);
-            }
-            TextCell {
-                ch,
-                start_col,
-                end_col: next_col.saturating_sub(1),
-            }
-        })
-        .collect()
-}
-
-fn cell_index_at_column(cells: &[TextCell], col: u16) -> Option<usize> {
-    cells
-        .iter()
-        .position(|cell| cell.start_col <= col && col <= cell.end_col)
-}
-
-fn byte_index_for_cell(row: &str, cell_idx: usize) -> usize {
-    row.char_indices()
-        .nth(cell_idx)
-        .map(|(idx, _)| idx)
-        .unwrap_or(row.len())
-}
-
-fn byte_index_after_cell(row: &str, cell_idx: usize) -> usize {
-    row.char_indices()
-        .nth(cell_idx.saturating_add(1))
-        .map(|(idx, _)| idx)
-        .unwrap_or(row.len())
-}
-
-fn url_span_at_column(cells: &[TextCell], clicked_idx: usize) -> Option<CellSpan> {
-    let mut start = 0;
-    while start < cells.len() {
-        if starts_with_chars(&cells[start..], "http://")
-            || starts_with_chars(&cells[start..], "https://")
-        {
-            let mut end = start;
-            while end + 1 < cells.len() && !cells[end + 1].ch.is_whitespace() {
-                end += 1;
-            }
-            if clicked_idx >= start && clicked_idx <= end {
-                let span = trim_url_edges(cells, CellSpan { start, end })?;
-                return span.contains(clicked_idx).then_some(span);
-            }
-            start = end + 1;
-        } else {
-            start += 1;
-        }
-    }
-    None
-}
-
-fn trim_url_edges(cells: &[TextCell], span: CellSpan) -> Option<CellSpan> {
-    let start = span.start;
-    let mut end = span.end;
-    while start <= end && should_trim_trailing_url_cell(cells, start, end) {
-        if end == 0 {
-            return None;
-        }
-        end -= 1;
-    }
-    (start <= end).then_some(CellSpan { start, end })
-}
-
-fn should_trim_trailing_url_cell(cells: &[TextCell], start: usize, end: usize) -> bool {
-    match cells[end].ch {
-        '"' | '\'' | '`' | '.' | ',' | ';' | ':' | '!' | '?' => true,
-        ')' => !trailing_url_closer_is_balanced(cells, start, end, '(', ')'),
-        ']' => !trailing_url_closer_is_balanced(cells, start, end, '[', ']'),
-        '}' => !trailing_url_closer_is_balanced(cells, start, end, '{', '}'),
-        _ => false,
-    }
-}
-
-fn trailing_url_closer_is_balanced(
-    cells: &[TextCell],
-    start: usize,
-    end: usize,
-    open: char,
-    close: char,
-) -> bool {
-    let mut balance = 0i32;
-    for cell in &cells[start..end] {
-        if cell.ch == open {
-            balance += 1;
-        } else if cell.ch == close {
-            balance -= 1;
-        }
-    }
-    balance > 0
-}
-
-fn quoted_path_span_at_column(cells: &[TextCell], clicked_idx: usize) -> Option<CellSpan> {
-    let clicked = cells.get(clicked_idx)?.ch;
-    if clicked == '"' || clicked == '\'' || clicked == '`' {
-        return None;
-    }
-
-    for quote in ['"', '\'', '`'] {
-        let mut start = None;
-        for (idx, cell) in cells.iter().copied().enumerate() {
-            let ch = cell.ch;
-            if ch != quote || is_escaped(cells, idx) {
-                continue;
-            }
-            if let Some(open) = start {
-                if clicked_idx > open
-                    && clicked_idx < idx
-                    && cells[open + 1..idx].iter().any(|cell| cell.ch == '/')
-                {
-                    return Some(CellSpan {
-                        start: open + 1,
-                        end: idx - 1,
-                    });
-                }
-                start = None;
-            } else {
-                start = Some(idx);
-            }
-        }
-    }
-    None
-}
-
-fn is_escaped(cells: &[TextCell], idx: usize) -> bool {
-    let mut slashes = 0;
-    let mut cursor = idx;
-    while cursor > 0 && cells[cursor - 1].ch == '\\' {
-        slashes += 1;
-        cursor -= 1;
-    }
-    slashes % 2 == 1
-}
-
-fn starts_with_chars(cells: &[TextCell], prefix: &str) -> bool {
-    prefix
-        .chars()
-        .enumerate()
-        .all(|(idx, expected)| cells.get(idx).is_some_and(|cell| cell.ch == expected))
-}
-
-fn is_word_separator(ch: char) -> bool {
-    ch.is_whitespace()
-        || matches!(
-            ch,
-            '|' | '(' | ')' | '[' | ']' | '{' | '}' | ',' | ';' | '!'
-        )
-}
-
-fn trim_token_edges(cells: &[TextCell], span: CellSpan) -> Option<CellSpan> {
-    let mut start = span.start;
-    let mut end = span.end;
-    while start <= end && is_leading_token_wrapper(cells[start].ch) {
-        start += 1;
-    }
-    if start < end && cells[end].ch == '$' && is_trailing_token_wrapper(cells[end - 1].ch) {
-        end -= 1;
-    }
-    while start <= end && is_trailing_token_wrapper(cells[end].ch) {
-        if end == 0 {
-            return None;
-        }
-        end -= 1;
-    }
-    (start <= end).then_some(CellSpan { start, end })
-}
-
-fn is_leading_token_wrapper(ch: char) -> bool {
-    matches!(ch, '(' | '[' | '{' | '<' | '"' | '\'' | '`')
-}
-
-fn is_trailing_token_wrapper(ch: char) -> bool {
-    matches!(
-        ch,
-        ')' | ']' | '}' | '>' | '"' | '\'' | '`' | '.' | ',' | ';' | ':' | '!' | '?'
-    )
-}
-
 // ---------------------------------------------------------------------------
 // Event handling
 // ---------------------------------------------------------------------------
@@ -2265,211 +1701,6 @@ mod tests {
     }
 
     #[test]
-    fn notification_context_formats_resolved_workspace_label() {
-        let state = app_with_workspaces(&["stale"]);
-        let root = state.workspaces[0].tabs[0].root_pane;
-
-        assert_eq!(
-            notification_context(&state.workspaces[0], "__herdr_projects__", 0, root),
-            "__herdr_projects__ · 1"
-        );
-    }
-
-    fn selected_word(row: &str, col: u16) -> Option<String> {
-        let (start, end) = word_bounds_at_column(row, col)?;
-        Some(text_in_cell_range(row, start, end))
-    }
-
-    fn selected_url<'a>(row: &'a str, click: &str) -> Option<&'a str> {
-        url_at_column(row, col_of(row, click))
-    }
-
-    fn text_in_cell_range(row: &str, start_col: u16, end_col: u16) -> String {
-        text_cells(row)
-            .into_iter()
-            .filter(|cell| cell.start_col >= start_col && cell.end_col <= end_col)
-            .map(|cell| cell.ch)
-            .collect()
-    }
-
-    fn col_of(row: &str, needle: &str) -> u16 {
-        let byte_idx = row
-            .find(needle)
-            .unwrap_or_else(|| panic!("{needle:?} not found in {row:?}"));
-        let prefix = &row[..byte_idx];
-        prefix
-            .chars()
-            .map(|ch| u16::from(crate::ghostty::unicode_codepoint_width(ch as u32)))
-            .sum()
-    }
-
-    fn assert_selects(row: &str, click: &str, expected: &str) {
-        assert_eq!(
-            selected_word(row, col_of(row, click)).as_deref(),
-            Some(expected),
-            "row={row:?}, click={click:?}"
-        );
-    }
-
-    fn assert_selects_nothing(row: &str, click: &str) {
-        assert_eq!(
-            selected_word(row, col_of(row, click)),
-            None,
-            "row={row:?}, click={click:?}"
-        );
-    }
-
-    #[test]
-    fn double_click_word_bounds_cover_terminal_text() {
-        let cases = [
-            (
-                "see https://example.com/a-b_c?q=x@y.",
-                "example.com",
-                "https://example.com/a-b_c?q=x@y",
-            ),
-            (
-                "open \"https://example.com/a,b;c?q=x\";",
-                "example.com",
-                "https://example.com/a,b;c?q=x",
-            ),
-            (
-                "see https://en.wikipedia.org/wiki/Foo_(bar_(baz)),",
-                "wikipedia",
-                "https://en.wikipedia.org/wiki/Foo_(bar_(baz))",
-            ),
-            (
-                "see https://example.com/a(b[c{d}e]f),",
-                "example.com",
-                "https://example.com/a(b[c{d}e]f)",
-            ),
-            (
-                "see (https://example.com/a(b(c)d)))",
-                "example.com",
-                "https://example.com/a(b(c)d)",
-            ),
-            (
-                "open /tmp/foo-bar/baz_qux/",
-                "foo-bar",
-                "/tmp/foo-bar/baz_qux/",
-            ),
-            (
-                "open ./src/app/actions.rs:795",
-                "actions",
-                "./src/app/actions.rs:795",
-            ),
-            (
-                "open ../herdr-worktrees/issue-1",
-                "herdr",
-                "../herdr-worktrees/issue-1",
-            ),
-            (
-                "edit src/app/actions.rs,then",
-                "actions",
-                "src/app/actions.rs",
-            ),
-            (
-                "cat \"/tmp/build output/log.txt\"",
-                "output",
-                "/tmp/build output/log.txt",
-            ),
-            (
-                "cat '/Users/me/Library/Application Support/app/config.json'",
-                "Support",
-                "/Users/me/Library/Application Support/app/config.json",
-            ),
-            ("echo 你好-world done", "好", "你好-world"),
-            ("先跑 cargo test", "cargo", "cargo"),
-            (
-                "export PATH=$HOME/.cargo/bin:$PATH",
-                "$HOME",
-                "PATH=$HOME/.cargo/bin:$PATH",
-            ),
-            (
-                "git checkout feature/foo-bar_baz",
-                "foo",
-                "feature/foo-bar_baz",
-            ),
-            ("refs #123 and @owner/name", "#123", "#123"),
-            ("refs #123 and @owner/name", "owner", "@owner/name"),
-            ("cargo test --package=herdr", "--package", "--package=herdr"),
-            (
-                "cargo test app::actions::tests",
-                "app::",
-                "app::actions::tests",
-            ),
-            (
-                "image ghcr.io/org/app:latest",
-                "ghcr",
-                "ghcr.io/org/app:latest",
-            ),
-            ("ERROR [worker-1] request_id=abc-123", "worker", "worker-1"),
-            (
-                "tmux|newhoo|fixhoo|newmoo|notification|window_bell|herdr",
-                "newhoo",
-                "newhoo",
-            ),
-            (
-                "render_status_line(app, area)",
-                "render",
-                "render_status_line",
-            ),
-            ("render_status_line(app, area)", "app", "app"),
-            ("render_status_line(app, area)", "area", "area"),
-            ("if !enabled {", "enabled", "enabled"),
-            ("println!(\"hi\")", "println", "println"),
-            ("( master)$", "master", "master"),
-            ("regex foo$", "foo", "foo$"),
-        ];
-
-        for (row, click, expected) in cases {
-            assert_selects(row, click, expected);
-        }
-
-        let row = "echo 你好-world done";
-        assert_eq!(
-            selected_word(row, col_of(row, "好") + 1).as_deref(),
-            Some("你好-world")
-        );
-    }
-
-    #[test]
-    fn double_click_word_bounds_ignore_delimiters() {
-        for (row, click) in [
-            (
-                "tmux|newhoo|fixhoo|newmoo|notification|window_bell|herdr",
-                "|",
-            ),
-            ("alpha,beta;gamma", ","),
-            ("alpha,beta;gamma", ";"),
-            ("render_status_line(app, area)", "("),
-            ("render_status_line(app, area)", ")"),
-            ("if !enabled {", "!"),
-            ("if !enabled {", "{"),
-            ("(done).", "("),
-            ("(done).", "."),
-        ] {
-            assert_selects_nothing(row, click);
-        }
-    }
-
-    #[test]
-    fn url_at_column_returns_safe_visible_url_only() {
-        assert_eq!(
-            selected_url("see https://example.com/a(b)c.", "example"),
-            Some("https://example.com/a(b)c")
-        );
-        assert_eq!(
-            selected_url("[docs](https://example.com/docs),", "example"),
-            Some("https://example.com/docs")
-        );
-        assert_eq!(
-            selected_url("[docs](https://example.com/docs)", "docs"),
-            None
-        );
-        assert_eq!(selected_url("open file:///tmp/report", "file"), None);
-    }
-
-    #[test]
     fn apply_workspace_git_statuses_updates_matching_workspace() {
         let mut state = app_with_workspaces(&["one", "two"]);
         let first_id = state.workspaces[0].id.clone();
@@ -3095,30 +2326,6 @@ mod tests {
     }
 
     #[test]
-    fn waiting_sound_plays_even_in_active_workspace() {
-        assert_eq!(
-            notification_sound_for_state_change(true, AgentState::Working, AgentState::Blocked),
-            Some(crate::sound::Sound::Request)
-        );
-    }
-
-    #[test]
-    fn done_sound_only_plays_in_background() {
-        assert_eq!(
-            notification_sound_for_state_change(false, AgentState::Working, AgentState::Idle),
-            Some(crate::sound::Sound::Done)
-        );
-        assert_eq!(
-            notification_sound_for_state_change(true, AgentState::Working, AgentState::Idle),
-            None
-        );
-        assert_eq!(
-            notification_sound_for_state_change(false, AgentState::Unknown, AgentState::Idle),
-            None
-        );
-    }
-
-    #[test]
     fn background_waiting_sets_attention_toast() {
         let mut state = app_with_workspaces(&["active", "background"]);
         state.active = Some(0);
@@ -3721,14 +2928,6 @@ mod tests {
         });
 
         assert!(state.toast.is_none());
-    }
-
-    #[test]
-    fn active_tab_suppression_preserves_unknown_focus_behavior() {
-        assert!(active_tab_suppresses_notifications(true, None));
-        assert!(active_tab_suppresses_notifications(true, Some(true)));
-        assert!(!active_tab_suppresses_notifications(true, Some(false)));
-        assert!(!active_tab_suppresses_notifications(false, None));
     }
 
     #[test]
