+++ src/server/notifications.rs
@@ -38,19 +38,20 @@ pub(crate) fn toast_message_from_state_change(
                     .terminals
                     .get(&pane.attached_terminal_id)
                     .and_then(|terminal| terminal.effective_agent_label())?;
-                let kind = app::actions::notification_toast_for_state_change_with_agent_labels(
-                    suppress_active_tab_notifications,
-                    prev_state,
-                    new_state,
-                    previous_agent_label,
-                    Some(agent_label),
-                )?;
+                let kind =
+                    app::notifications::notification_toast_for_state_change_with_agent_labels(
+                        suppress_active_tab_notifications,
+                        prev_state,
+                        new_state,
+                        previous_agent_label,
+                        Some(agent_label),
+                    )?;
                 let workspace_label = ws.display_name_from(&state.terminals, terminal_runtimes);
                 Some(format!(
                     "{} {}: {}",
                     agent_label,
                     toast_event_text(kind),
-                    app::actions::notification_context(ws, &workspace_label, ws_idx, pane_id)
+                    app::notifications::notification_context(ws, &workspace_label, ws_idx, pane_id)
                 ))
             })
         })
